Split the CRD overlap validation from csv-merger to a new tool
The CRD overlap validation is used only in build-manifest, and not in
any other place. It should not be part of the csv-merger, and not part
of the operator image.

This PR take this functionality out of the csv-merger, and introduces
the new crd-overlap-validator tool, to be used from the build-manifests
script.
