+++ tools/csv-merger/csv-merger.go
@@ -24,11 +24,8 @@ import (
 	"errors"
 	"flag"
 	"fmt"
-	"io"
 	"log"
 	"os"
-	"path/filepath"
-	"regexp"
 	"slices"
 	"sort"
 	"strings"
@@ -38,8 +35,6 @@ import (
 	"github.com/ghodss/yaml"
 	csvv1alpha1 "github.com/operator-framework/api/pkg/operators/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apiextensions-apiserver/pkg/apis/apiextensions"
-	"k8s.io/apimachinery/pkg/util/sets"
 
 	"github.com/kubevirt/hyperconverged-cluster-operator/pkg/components"
 	hcoutil "github.com/kubevirt/hyperconverged-cluster-operator/pkg/util"
@@ -114,7 +109,6 @@ var (
 	relatedImagesList = flag.String("related-images-list", "",
 		"Comma separated list of all the images referred in the CSV (just the image pull URLs or eventually a set of 'image|name' collations)")
 	ignoreComponentsRelatedImages = flag.Bool("ignore-component-related-image", false, "Ignore relatedImages from components CSVs")
-	crdDir                        = flag.String("crds-dir", "", "the directory containing the CRDs for apigroup validation. The validation will be performed if and only if the value is non-empty.")
 	hcoKvIoVersion                = flag.String("hco-kv-io-version", "", "KubeVirt version")
 	kubevirtVersion               = flag.String("kubevirt-version", "", "Kubevirt operator version")
 	kvVirtLauncherOSVersion       = flag.String("virt-launcher-os-version", "", "Virt launcher OS version")
@@ -135,100 +129,6 @@ var (
 	envVars EnvVarFlags
 )
 
-func IOReadDir(root string) ([]string, error) {
-	var files []string
-	fileInfo, err := os.ReadDir(root)
-	if err != nil {
-		return files, err
-	}
-
-	for _, file := range fileInfo {
-		files = append(files, filepath.Join(root, file.Name()))
-	}
-	return files, nil
-}
-
-func validateNoAPIOverlap(crdDir string) error {
-	crdFiles, err := IOReadDir(crdDir)
-	if err != nil {
-		return err
-	}
-
-	// crdMap is populated with operator names as keys and a slice of associated api groups as values.
-	crdMap := getCrdMap(crdFiles)
-
-	overlapsMap := detectAPIOverlap(crdMap)
-
-	return checkAPIOverlapMap(overlapsMap)
-}
-
-func checkAPIOverlapMap(overlapsMap map[string]sets.Set[string]) error {
-	// if at least one overlap found - emit an error.
-	if len(overlapsMap) != 0 {
-		var sb strings.Builder
-		// WriteString always returns error=nil. no point to check it.
-		_, _ = sb.WriteString("ERROR: Overlapping API Groups were found between different operators.\n")
-		for apiGroup := range overlapsMap {
-			_, _ = sb.WriteString(fmt.Sprintf("The API Group %s is being used by these operators: %s\n", apiGroup, strings.Join(overlapsMap[apiGroup].UnsortedList(), ", ")))
-		}
-		return errors.New(sb.String())
-	}
-	return nil
-}
-
-func detectAPIOverlap(crdMap map[string][]string) map[string]sets.Set[string] {
-	// overlapsMap is populated with collisions found - API Groups as keys,
-	// and slice containing operators using them, as values.
-	overlapsMap := make(map[string]sets.Set[string])
-	for operator, groups := range crdMap {
-		for _, apiGroup := range groups {
-			compareMapWithEntry(crdMap, operator, apiGroup, overlapsMap)
-		}
-	}
-	return overlapsMap
-}
-
-func compareMapWithEntry(crdMap map[string][]string, operator string, apigroup string, overlapsMap map[string]sets.Set[string]) {
-	for comparedOperator := range crdMap {
-		if operator == comparedOperator { // don't check self
-			continue
-		}
-
-		if slices.Contains(crdMap[comparedOperator], apigroup) {
-			overlapsMap[apigroup].Insert(operator)
-			overlapsMap[apigroup].Insert(comparedOperator)
-		}
-	}
-}
-
-func getCrdMap(crdFiles []string) map[string][]string {
-	crdMap := make(map[string][]string)
-
-	for _, crdFilePath := range crdFiles {
-		file, err := os.Open(crdFilePath)
-		panicOnError(err)
-
-		content, err := io.ReadAll(file)
-		panicOnError(err)
-
-		err = file.Close()
-		panicOnError(err)
-
-		crdFileName := filepath.Base(crdFilePath)
-		reg := regexp.MustCompile(`([^\d]+)`)
-		operator := reg.FindString(crdFileName)
-
-		var crd apiextensions.CustomResourceDefinition
-		err = yaml.Unmarshal(content, &crd)
-		panicOnError(err)
-
-		if !slices.Contains(crdMap[operator], crd.Spec.Group) {
-			crdMap[operator] = append(crdMap[operator], crd.Spec.Group)
-		}
-	}
-	return crdMap
-}
-
 func main() {
 	flag.Var(&envVars, "env-var", "HCO environment variable (key=value), may be used multiple times")
 
@@ -236,11 +136,6 @@ func main() {
 
 	if webhookImage == nil || *webhookImage == "" {
 		*webhookImage = *operatorImage
-	}
-
-	if *crdDir != "" {
-		panicOnError(validateNoAPIOverlap(*crdDir))
-		os.Exit(0)
 	}
 
 	if *enableUniqueSemver && *olmSkipRange != "" {
